Fixed crash when attempting to reading tokens past final in list
I wrapped reading a token in a function that will return an error and
report via JsonDiag if someone tries to access tokens past the end of
the list.
It seems that my value_separator code intruduced this bug. Instead of
fixing that one case, and hoping it doesn't show up elsewhere, I decided
to prevent it with the wrapper function.